Renamed package to dev.masa
