+++ src/main/java/dev/masa/masuitecore/bungee/listeners/CoreMessageListener.java
@@ -1,0 +1,25 @@
+package dev.masa.masuitecore.bungee.listeners;
+
+import net.md_5.bungee.api.event.PluginMessageEvent;
+import net.md_5.bungee.api.plugin.Listener;
+import net.md_5.bungee.event.EventHandler;
+
+import java.io.ByteArrayInputStream;
+import java.io.DataInputStream;
+import java.io.IOException;
+
+public class CoreMessageListener implements Listener {
+
+    @EventHandler
+    public void onPluginMessage(PluginMessageEvent e) throws IOException {
+        if (!e.getTag().equals("BungeeCord")) {
+            return;
+        }
+        DataInputStream in = new DataInputStream(new ByteArrayInputStream(e.getData()));
+        String subchannel = in.readUTF();
+        if (subchannel.equals("MaSuiteDebugMessage")) {
+            System.out.println(in.readUTF());
+        }
+    }
+}
+
