+++ src/main/java/dev/masa/masuitecore/core/adapters/SpongeAdapter.java
@@ -1,0 +1,5 @@
+package dev.masa.masuitecore.core.adapters;
+
+public class SpongeAdapter {
+
+}
