+++ src/main/java/dev/masa/masuitecore/core/api/MaSuiteCoreAPI.java
@@ -1,0 +1,15 @@
+package dev.masa.masuitecore.core.api;
+
+import dev.masa.masuitecore.core.services.PlayerService;
+import dev.masa.masuitecore.bungee.MaSuiteCore;
+
+public class MaSuiteCoreAPI {
+
+    public static MaSuiteCore getCore() {
+        return MaSuiteCore.getInstance();
+    }
+
+    public PlayerService getPlayerService() {
+        return MaSuiteCore.getInstance().playerService;
+    }
+}
