--- src/main/java/fi/matiaspaavilainen/masuitecore/core/adapters/SpongeAdapter.java
@@ -1,5 +1,0 @@
-package fi.matiaspaavilainen.masuitecore.core.adapters;
-
-public class SpongeAdapter {
-
-}
