--- src/main/java/fi/matiaspaavilainen/masuitecore/core/api/MaSuiteCoreAPI.java
@@ -1,15 +1,0 @@
-package fi.matiaspaavilainen.masuitecore.core.api;
-
-import fi.matiaspaavilainen.masuitecore.bungee.MaSuiteCore;
-import fi.matiaspaavilainen.masuitecore.core.services.PlayerService;
-
-public class MaSuiteCoreAPI {
-
-    public static MaSuiteCore getCore() {
-        return MaSuiteCore.getInstance();
-    }
-
-    public PlayerService getPlayerService() {
-        return MaSuiteCore.getInstance().playerService;
-    }
-}
